requirements.txt
# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Vector database
qdrant-client==1.11.3

# Redis
redis[hiredis]==5.0.1

# Object storage (MinIO/S3)
minio==7.2.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pytesseract==0.3.10
Pillow>=10.2.0

# ML and embeddings
sentence-transformers==2.2.2
torch>=2.1.1
numpy>=1.24.4

# AI APIs
openai>=1.3.7
httpx>=0.24.1

# WebSocket support
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1 